+++ .dialyzer_ignore.exs
@@ -1,0 +1,8 @@
+[
+  # Ignore warnings caused by the following problem:
+  # https://github.com/elixir-lang/elixir/issues/6738
+  {"lib/geometry.ex", :pattern_match},
+  {"lib/geometry.ex", :no_return},
+  {"lib/geometry/decoder/wkt.ex", :pattern_match},
+  {"lib/geometry/decoder/wkt.ex", :unused_fun}
+]